+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -161,5 +161,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-28T12:57:47.340Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-28T15:44:09.367Z")
     r"""The date and time the resource was last updated"""